+++ CHANGELOG.md
@@ -77,6 +77,8 @@ Our versioning strategy is as follows:
 * `[core]` SXA Form can't fire CloudSDK events due to initialization error ([#63](https://github.com/Sitecore/content-sdk/pull/63)):
   * Form utilities have been moved from `@sitecore-content-sdk/core/form` to the root of `@sitecore-content-sdk/core`. Update your imports to reflect this change if you are referencing these utilities.
 * `[nextjs]` Update React to version 19 and Next JS to version 15 ([#76](https://github.com/Sitecore/content-sdk/pull/76))
+* `[nextjs]` Component-level `getServerSideProps` and `getStaticProps` methods have been replaced by a single `getComponentServerProps` method for simplicity.
+  * In case a separate logic is needed depending on SSR/SSG context, a `isServerSidePropsContext` helper method from `@sitecore-content-sdk/nextjs/utils` can now be used.
 
 ### 🐛 Bug Fixes